Fixes to movement algorithm

+++ jgsf/src/main/java/com/jenjinstudios/world/WorldObject.java
@@ -16,7 +16,7 @@ public class WorldObject
 	/** The vector2D in the world at which the object is located. */
 	private Vector2D vector2D;
 	/** The direction in which this object is facing. */
-	private float direction;
+	private double direction;
 	/** The ID number of this object. */
 	private int id = Integer.MIN_VALUE;
 	/** The world in which this object exists. */
@@ -46,7 +46,7 @@ public class WorldObject
 	 *
 	 * @return The direction in which this object is facing.
 	 */
-	public float getDirection()
+	public double getDirection()
 	{
 		return direction;
 	}
@@ -56,7 +56,7 @@ public class WorldObject
 	 *
 	 * @param direction The new direction for this object to face.
 	 */
-	public void setDirection(float direction)
+	public void setDirection(double direction)
 	{
 		this.direction = direction;
 	}
